restructred example folder, add vectors / gravity
