+++ example/ameiva/falling_stars/index.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>Document</title>
+    <style>
+        .main{
+            background-color:black;
+            border: 1px solid blueviolet;
+        }
+    </style>
+    <script  src="main.dart.js"></script>
+</head>
+<body>
+    <!-- <div id="fpsDisplay" style="position: absolute; top: 70px; left: 10px;"></div> -->
+</body>
+</html>
