--- example/mainLLL.dart
@@ -1,6 +1,0 @@
-// import 'ameiva_falling_stars.dart';
-// void main() => fallingStars();
-
-import 'vectors.dart';
-
-main() => vectorSetup();
