+++ example/vectors/helpers.dart
